Rename DefaultContentType to SimpleContentType and hide it from the public API
Moves the record out of the ContentType interface (where interface nesting
forced it public) to a top-level package-private sibling in LogOutput.java,
matching the existing pattern used for FormatterEncoder/StdOutOutput/etc.
Public javadoc no longer names the type at all, since it's just an internal
implementation detail of ContentType.of(...).
