+++ core/src/test/java/io/jstach/rainbowgum/ContentTypeTest.java
@@ -1,4 +1,4 @@
-package io.jstach.rainbowgum.output;
+package io.jstach.rainbowgum;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertSame;
@@ -8,9 +8,13 @@ import java.nio.charset.StandardCharsets;
 import org.junit.jupiter.api.Test;
 
 import io.jstach.rainbowgum.LogOutput.ContentType;
-import io.jstach.rainbowgum.LogOutput.ContentType.DefaultContentType;
 import io.jstach.rainbowgum.LogOutput.ContentType.StandardContentType;
 
+/*
+ * Moved into io.jstach.rainbowgum (from io.jstach.rainbowgum.output) so this can
+ * construct SimpleContentType directly - it's package-private now (see the "minimize
+ * public API" note), not something outside this package should ever need to touch.
+ */
 class ContentTypeTest {
 
 	@Test
@@ -30,9 +34,9 @@ class ContentTypeTest {
 	}
 
 	@Test
-	void ofCreatesADefaultContentTypeWhenNothingMatches() {
+	void ofCreatesASimpleContentTypeWhenNothingMatches() {
 		var contentType = ContentType.of("text/plain", StandardCharsets.ISO_8859_1);
-		assertEquals(new DefaultContentType("text/plain", StandardCharsets.ISO_8859_1), contentType);
+		assertEquals(new SimpleContentType("text/plain", StandardCharsets.ISO_8859_1), contentType);
 		assertEquals("text/plain", contentType.contentType());
 		assertEquals(StandardCharsets.ISO_8859_1, contentType.charsetOrNull());
 	}
@@ -40,7 +44,7 @@ class ContentTypeTest {
 	@Test
 	void ofDoesNotMatchApplicationJsonWithAWrongCharset() {
 		var contentType = ContentType.of("application/json", StandardCharsets.ISO_8859_1);
-		assertEquals(new DefaultContentType("application/json", StandardCharsets.ISO_8859_1), contentType);
+		assertEquals(new SimpleContentType("application/json", StandardCharsets.ISO_8859_1), contentType);
 	}
 
 }
