+++ core/src/test/java/io/jstach/rainbowgum/LogEncoderCharsetTest.java
@@ -12,7 +12,6 @@ import java.util.List;
 import org.junit.jupiter.api.Test;
 
 import io.jstach.rainbowgum.LogOutput.ContentType;
-import io.jstach.rainbowgum.LogOutput.ContentType.DefaultContentType;
 import io.jstach.rainbowgum.LogOutput.WriteMethod;
 
 /*
@@ -63,7 +62,7 @@ class LogEncoderCharsetTest {
 	@Test
 	void ofContentTypeUsesTheGivenContentTypeAndItsCharset() {
 		var output = new CapturingOutput(WriteMethod.BYTES);
-		var contentType = new DefaultContentType("text/csv", StandardCharsets.ISO_8859_1);
+		var contentType = new SimpleContentType("text/csv", StandardCharsets.ISO_8859_1);
 		var config = LogConfig.builder().build();
 		var gum = RainbowGum.builder(config)
 			.route(r -> r.appender("list",
@@ -79,7 +78,7 @@ class LogEncoderCharsetTest {
 	@Test
 	void ofContentTypeWithNoCharsetDefaultsToUtf8ForEncoding() {
 		var output = new CapturingOutput(WriteMethod.BYTES);
-		var contentType = new DefaultContentType("text/csv", null);
+		var contentType = new SimpleContentType("text/csv", null);
 		var config = LogConfig.builder().build();
 		var gum = RainbowGum.builder(config)
 			.route(r -> r.appender("list",
